add SoftDelete decorator
close https://rvohealth.atlassian.net/browse/PDTC-5520

+++ src/dream/internal/destroyDream.ts
@@ -2,6 +2,7 @@ import Dream from '../../dream'
 import DreamTransaction from '../transaction'
 import runHooksFor from './runHooksFor'
 import safelyRunCommitHooks from './safelyRunCommitHooks'
+import softDeleteDream from './softDeleteDream'
 
 /**
  * @internal
@@ -16,14 +17,14 @@ import safelyRunCommitHooks from './safelyRunCommitHooks'
 export default async function destroyDream<I extends Dream>(
   dream: I,
   txn: DreamTransaction<I> | null = null,
-  { skipHooks = false }: { skipHooks?: boolean } = {}
+  { skipHooks = false, reallyDestroy = false }: { skipHooks?: boolean; reallyDestroy?: boolean } = {}
 ): Promise<I> {
   if (txn) {
-    return await destroyDreamWithTransaction(dream, txn, { skipHooks })
+    return await destroyDreamWithTransaction(dream, txn, { skipHooks, reallyDestroy })
   } else {
     const dreamClass = dream.constructor as typeof Dream
     return await dreamClass.transaction(
-      async txn => await destroyDreamWithTransaction<I>(dream, txn, { skipHooks })
+      async txn => await destroyDreamWithTransaction<I>(dream, txn, { skipHooks, reallyDestroy })
     )
   }
 }
@@ -38,17 +39,18 @@ export default async function destroyDream<I extends Dream>(
 async function destroyDreamWithTransaction<I extends Dream>(
   dream: I,
   txn: DreamTransaction<I>,
-  { skipHooks }: { skipHooks: boolean }
+  { skipHooks, reallyDestroy }: { skipHooks: boolean; reallyDestroy: boolean }
 ): Promise<I> {
+  await destroyAssociatedRecords(dream, txn, { skipHooks, reallyDestroy })
+
   if (!skipHooks) {
     await runHooksFor('beforeDestroy', dream, true, null, txn)
   }
 
-  await destroyAssociatedRecords(dream, txn, { skipHooks })
-  await maybeDestroyDream(dream, txn)
+  await maybeDestroyDream(dream, txn, reallyDestroy)
 
   if (!skipHooks) {
-    await runHooksFor('afterDestroy', dream, true, null, txn)
+    await runHooksFor('afterDestroy', dream, true, null, txn || undefined)
     await safelyRunCommitHooks(dream, 'afterDestroyCommit', true, null, txn)
   }
 
@@ -61,7 +63,17 @@ async function destroyDreamWithTransaction<I extends Dream>(
  * Destroys the dream iff it was not blocked from
  * deleting by one of the beforeDestroy model hooks
  */
-async function maybeDestroyDream<I extends Dream>(dream: I, txn: DreamTransaction<I>) {
+async function maybeDestroyDream<I extends Dream>(
+  dream: I,
+  txn: DreamTransaction<I>,
+  reallyDestroy: boolean
+) {
+  const dreamClass = dream.constructor as typeof Dream
+
+  if (dreamClass['softDelete'] && !reallyDestroy) {
+    await softDeleteDream(dream, txn)
+  }
+
   if (dream['_preventDeletion']) {
     dream.unpreventDeletion()
   } else {
@@ -81,11 +93,15 @@ async function maybeDestroyDream<I extends Dream>(dream: I, txn: DreamTransactio
 async function destroyAssociatedRecords<I extends Dream>(
   dream: I,
   txn: DreamTransaction<I>,
-  { skipHooks }: { skipHooks: boolean }
+  { skipHooks, reallyDestroy }: { skipHooks: boolean; reallyDestroy: boolean }
 ) {
   const dreamClass = dream.constructor as typeof Dream
 
   for (const associationName of dreamClass['dependentDestroyAssociationNames']()) {
-    await dream.txn(txn).destroyAssociation(associationName as any, { skipHooks })
+    if (reallyDestroy) {
+      await dream.txn(txn).reallyDestroyAssociation(associationName as any, { skipHooks })
+    } else {
+      await dream.txn(txn).destroyAssociation(associationName as any, { skipHooks })
+    }
   }
 }
